chrono changes
chrono v0.3 -> v0.4

Main change is in its ISO week number API, which prompted me to take a
hard look at how I was using the ISO week number in the first place.
I've decided that it was not really correct, and so I've changed / fixed
it to something better.

Also made the automanage code hopefully clearer to read, and made its
output explain why it's deleting a snapshot.